Rename test files.
- Rename test.js to test-node.js. Aligns with test-karma.js.
- Rename test-common.js to test.js.
